use client id option in ceph commands
By default ceph cli is run with 'admin' client id. But the admin
keyring file is not installed on all nodes. In this case instead of
real data the collected output looks like below:

  #==[ Command ]======================================#
  # /usr/bin/ceph --connect-timeout=5 -s
  2019-03-13 15:18:30.626774 7f731b676700 -1 auth: unable to find a keyring on /etc/ceph/ceph.client.admin.keyring,/etc/ceph/ceph.keyring,/etc/ceph/keyring,/etc/ceph/keyring.bin,: (2) No such file or directory
  2019-03-13 15:18:30.626781 7f731b676700 -1 monclient: ERROR: missing keyring, cannot use cephx for authentication
  2019-03-13 15:18:30.626782 7f731b676700  0 librados: client.admin initialization error (2) No such file or directory
  [errno 2] error connecting to the cluster

Still a keyring file for some other client id (e.g. storage) is
usually available and ceph command run with this id would succeed.

So look for the available keyring file and set client id accordingly.

Signed-off-by: Mykola Golub <mgolub@suse.com>