Don't try to use non-admin keyring for getting ceph status
In order to invoke all the `ceph` commands to get cluster
status, the current implementation tries to use whatever
keyring it can find in /etc/ceph/ if the admin keyring isn't
present.  Now that everything is running in containers, I
think it's less likely that any alternative keys will be
available to use on the host, so I've removed this
functionality.  This reverts 59843bd.

Note:
- The alternate keyring functionality wasn't working anyway
  with `cephadm shell`.  If we wanted to keep it, we'd have
  to add the --keyring paramenter to make cephadm use the
  right keyring, as well as setting $ID.
- Even without this functionality, supportconfig will still
  capture individual daemon state via `cephadm enter`.  It
  just won't get general cluster status.
- There will always be at least one node with an admin
  keyring (the bootstrap node), so cluster status can always
  be obtained by running supportconfig on that node.

Signed-off-by: Tim Serong <tserong@suse.com>