ses: Eliminate implicit dependency on "which"
The "which" command is not part of "coreutils" - instead, it is shipped
in a separate RPM, which might not be installed unless it's declared as
a dependency in the supportutils-plugin-ses.spec file.

One option would be to make supportutils-plugin-ses depend on the
"which" RPM, but it turns out that "type" (a shell builtin mandated by
POSIX) works just as well.

Signed-off-by: Nathan Cutler <ncutler@suse.com>