Prepare for merge of Rook collection functionality
This merely paves the way for easy introduction of Rook log collection
functionality. Basically, it makes the 'ses' plugin a small file that
calls out to larger plugin 'resource' files. The new idea is to make
more use of the /usr/lib/supportconfig/resources directory for
containing resources, especially since there is a lot of code for
collecting Ceph logs. Most of the code will live in resources and be
called/loaded as-needed.

This will mean that the build service package needs to be built slightly
differently. Instead of placing the 'ses-rpm-list' file into the plugin
resources directory, the whole new 'ses-resources' directory here in
source will need to be installed to /usr/lib/supportconfig/resources/ses

I teased out the preexisting validate_rpm_if_installed() function into a
helpers.sh in the ses-resources/ dir to prove out the above change.
Additionally, most of the current validation code (almost all) has been
moved to the ses-resources/cephadm script.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>